refactor: move some logic to a module
