+++ src/components/Game.tsx
@@ -1,11 +1,11 @@
-import { useState, useCallback, useMemo } from 'react';
+import { useState, useMemo } from 'react';
 import { BoardSquares } from '../types/BoardSquares';
 import { History } from '../types/History';
 import { SortOrder } from '../types/SortOrder';
 import { Board } from './Board';
-import { conditionsToWin } from '../contracts';
 import ToggleSortOrderButton from './ToggleSortOrderButton';
 import styles from './Game.module.css';
+import { calculateMoveLocation, calculateWinner } from '../lib/game';
 
 export const Game = () => {
   const [history, setHistory] = useState<History>([
@@ -14,27 +14,10 @@ export const Game = () => {
   const [xIsNext, setXISNext] = useState(true);
   const [stepNumber, setStepNumber] = useState(0);
   const [sortOrder, setSortOrder] = useState<SortOrder>('ascending');
-
-  const calculateWinnersLine = useCallback((squares: BoardSquares) => {
-    const lineCausedOfWin = conditionsToWin.find((wayOfWinning) => {
-      const [a, b, c] = wayOfWinning;
-      return (
-        squares[a] !== null &&
-        squares[a] === squares[b] &&
-        squares[a] === squares[c]
-      );
-    });
-
-    return lineCausedOfWin ?? null;
-  }, []);
-
-  const calculateWinner = useCallback(
-    (squares: BoardSquares) => {
-      const lineCausedOfWin = calculateWinnersLine(squares);
-
-      return lineCausedOfWin !== null ? squares[lineCausedOfWin[0]] : null;
-    },
-    [calculateWinnersLine],
+  const current = history[stepNumber];
+  const winner = useMemo(
+    () => calculateWinner(current.squares),
+    [current.squares],
   );
 
   const handleClick = (i: number) => {
@@ -42,7 +25,6 @@ export const Game = () => {
 
     const newSquares = current.squares.slice() as BoardSquares;
     newSquares[i] = xIsNext ? 'X' : 'O';
-    // The second parameter of Array.prototype.slice() doesn't include its element, so add 1 to stepNumber.
     const newHistory = history.slice(0, stepNumber + 1);
 
     setHistory(newHistory.concat([{ squares: newSquares }]));
@@ -55,41 +37,10 @@ export const Game = () => {
     setXISNext(stepNumber % 2 === 0);
   };
 
-  const current = history[stepNumber];
-  const winner = useMemo(
-    () => calculateWinner(current.squares),
-    [calculateWinner, current.squares],
-  );
-
-  const getMoveLocation = (step: number) => {
-    const previous = history[step - 1].squares;
-    const target = history[step].squares;
-    const movedIndex = previous.findIndex(
-      (value, index) => value !== target[index],
-    );
-    const locations = [
-      [1, 1],
-      [2, 1],
-      [3, 1],
-      [1, 2],
-      [2, 2],
-      [3, 2],
-      [1, 3],
-      [2, 3],
-      [3, 3],
-    ] as const;
-
-    return movedIndex === -1 ? null : locations[movedIndex];
-  };
-
   return (
     <div className={styles.game}>
       <div>
-        <Board
-          squares={current.squares}
-          handleClick={handleClick}
-          highlights={calculateWinnersLine(current.squares)}
-        />
+        <Board squares={current.squares} handleClick={handleClick} />
       </div>
       <div className={styles.gameInfo}>
         <div>
@@ -119,7 +70,7 @@ export const Game = () => {
                   onClick={() => jumpTo(move)}
                 >
                   {move !== 0 ? `Go to move #${move}` : 'Go to game start'}
-                  {move !== 0 && ` (${getMoveLocation(move)})`}
+                  {move !== 0 && ` (${calculateMoveLocation(move, history)})`}
                 </button>
               </li>
             ))}
